fixes from Ei-ji Nakama
git-svn-id: https://svn.r-project.org/R/trunk@34680 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/gram.c
@@ -3538,7 +3538,7 @@ static int StringValue(int c)
                     if ((c = xxgetc()) != '}')
                         error(_("invalid \\u{xxxx} sequence"));
                 res = wcrtomb(buff, val, NULL); /* should always be valid */
-                if (res <= 0)
+                if ((int)res <= 0)
                     error(_("invalid \\uxxxx sequence"));
                 for (i = 0; i < res - 1; i++)
                     YYTEXT_PUSH(buff[i], yyp);
@@ -3576,7 +3576,7 @@ static int StringValue(int c)
                     if ((c = xxgetc()) != '}')
                         error(_("invalid \\U{xxxxxxxx} sequence"));
                 res = wcrtomb(buff, val, NULL); /* should always be valid */
-                if (res <= 0)
+                if ((int)res <= 0)
                     error(("invalid \\Uxxxxxxxx sequence"));
                 for (i = 0; i < res - 1; i++)
                     YYTEXT_PUSH(buff[i], yyp);
